Modulo_3/Aulas/treino_em_aula-semana09/py-mongo-flask/pyproject.toml
[tool.poetry]
name = "py-mongo-flask"
version = "0.1.0"
description = ""
authors = ["ViniciusPosssatto <101053966+ViniciusPosssatto@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
python-dotenv = "^0.21.0"
flasgger = "^0.9.5"
bcrypt = "^4.0.1"
requests = "^2.28.1"
PyJWT = "^2.6.0"
Flask-Cors = "^3.0.10"
Flask-PyMongo = "^2.3.0"
pymongo = "^4.3.2"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
